{{cookiecutter.repository_name}}/requirements.txt
# Django 
Django==2.2.2
django-allauth==0.39.1
django-configurations==2.1
django-filter==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
django-storages==1.7.1

# GraphQL
graphene==2.1.7
graphene-django==2.4.0

# Testing
coverage==4.5.3
factory-boy==2.12.0

# AWS
boto3==1.9.176

# Database
psycopg2-binary==2.8.3

# Misc
black==19.3b0

# Server
gunicorn==19.9.0
